Added udev restart, changed img format to jpg+raw, small fixes
The udev service needs to be restarted in order to have write
access to the root partition. Not noticed during testing, since
the service was restarted a couple of times.

The ":" in image file names had to be changed to _, since the data
cannot be copied to a USB stick with a vfat file system otherwise.